+++ src/components/Feeds/Feed/FeedItem/LikedArea.styled.js
